+++ R/zzz.R
@@ -1,2 +1,2 @@
-#' @useDynLib microdatasus microdatasus_dbc2dbf microdatasus_read_dbc
+#' @useDynLib microdatasus microdatasus_dbc2dbf microdatasus_dbc_info microdatasus_read_dbc
 NULL